tools: setup-conf: Change the UUID gen logics
Add EMMC_TARGET=/dev/mmcblk2. Its value will be used as a pattern for
the uuidgen.
The machine internal storage is emmc and its device node is /dev/mmcblk2.
The same pattern is getting used by the cl-deploy.mod script at the end
of the cl-deploy process on the device. The device uuid must be the same
as the one in the Yocto build environment. Having the same uuid for
development and at the device allows creating mender artifacts using the
Yocto environment.

Signed-off-by: Valentin Raevsky <valentin@compulab.co.il>